refactor: move category icon selector into shared control
